+++ piwigo/Albums/Discover/TaggedImagesViewController.m
@@ -439,7 +439,7 @@
         
         // Image Id?
         NSInteger imageId = [[userInfo objectForKey:@"imageId"] integerValue];
-        NSLog(@"=> removeImage %ld to Category %ld", imageId, catId);
+        NSLog(@"=> removeImage %ld to Category %ld", (long)imageId, (long)catId);
         
         // Store current image list
         NSArray *oldImageList = self.albumData.images;
